requirements.txt
hydra-core>=1.3.2
omegaconf>=2.3.0
numpy>=1.26.0
wandb>=0.17.0
pytest>=8.0.0
